Removed various instances of np.int

+++ ase/lattice/bravais.py
@@ -258,7 +258,7 @@ class Bravais:
         self.natoms = self.calc_num_atoms()
         self.nput = 0
         self.atoms = np.zeros((self.natoms, 3), np.float)
-        self.elements = np.zeros(self.natoms, np.int)
+        self.elements = np.zeros(self.natoms, int)
         self.farpoint = sum(self.directions)
         # printprogress = self.debug and (len(self.atoms) > 250)
         # Find the radius of the sphere containing the whole system
